feat: integration test and minor changes
